Refactor navigation items and ux cleanup
