+++ app/Domits/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { AuthProvider } from './src/context/AuthContext'; // Import AuthProvider from the context
-import AppNavigation from './src/navigation/appNavigation'; // Adjust the import if necessary
+import BottomTabNavigator from './src/navigation/appNavigation'; // Adjust the import if necessary
 
 function App() {
     return (
         <AuthProvider>
             <NavigationContainer>
-                <AppNavigation />
+                <BottomTabNavigator />
             </NavigationContainer>
         </AuthProvider>
     );
