+++ app/Domits/src/navigation/appNavigation.js
@@ -1,120 +1,17 @@
 import React from 'react';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
-import HomeScreen from '../screens/homeScreen';
-import Messages from '../screens/messages';
-import AccountPage from '../screens/account';
-import GuestDashboard from '../screens/guestdashboard/guestDashboard';
-import Scan from '../header/scan';
-import Pay from '../header/pay';
-import Bookings from '../header/bookings';
-import Pocket from '../header/pocket';
 import {View, Text} from 'react-native';
-import profile from '../screens/guestdashboard/GuestProfileTab';
-import paymentMethods from '../screens/guestdashboard/GuestPaymentsTab';
-import reviews from '../screens/guestdashboard/GuestReviewsTab';
-import settings from '../screens/guestdashboard/GuestSettingsTab';
-import helpAndFeedback from '../screens/guestdashboard/GuestHelpAndFeedbackTab';
-// import HostHomepage from '../screens/hostdashboard/hostDashboard';
-import HostDashboard from '../screens/hostdashboard/HostDashboard';
-import HostCalendar from '../screens/hostdashboard/HostCalendarTab';
-import HostReviews from '../screens/hostdashboard/HostReviewsTab';
-import HostProfile from '../screens/hostdashboard/HostProfileTab';
-import HostPayments from '../screens/hostdashboard/HostPaymentsTab';
-import HostListings from '../screens/hostdashboard/HostListingsTab';
-import OnboardingHost from '../screens/apphostonboarding/OnboardingHost';
-import HostSettings from '../screens/hostdashboard/HostSettingsTab';
-import Landing from '../screens/Landing';
-import emailSettings from '../screens/hostdashboard/emailSettings';
-// import HostRevenue from '../screens/hostdashboard/HostRevenueTab';
-import Detailpage from '../screens/detailpage';
-import onBoarding1 from '../screens/bookingprocess/bookingProcess';
-import personalDetailsForm from '../screens/bookingprocess/personalDetailsForm';
-import finalBookingOverview from '../screens/bookingprocess/finalBookingOverview';
-import simulateStripe from '../screens/bookingprocess/simulateStripe';
-import paymentAccepted from '../screens/bookingprocess/paymentAccepted';
-import paymentDeclined from '../screens/bookingprocess/paymentDeclined';
-import bookedAccommodation from '../screens/bookingprocess/bookedAccommodation';
-import ListProperty from '../screens/hostdashboard/HostListingsTab';
-// import LocationFillIn from '../screens/hostonboarding/LocationFillIn';
-// import selectAmenities from '../screens/hostonboarding/selectAmenities';
-// import PriceProperty from '../screens/hostonboarding/PriceProperty';
-import SignupScreen from '../screens/login/signup';
-import LoginScreen from '../screens/login/loginScreen';
-import Register from '../screens/login/register';
-import ConfirmMail from '../screens/login/confirmMail';
-import HostDetailPage from '../screens/hostdashboard/HostDetailPage';
-// import ReviewAndSubmitScreen from '../screens/hostonboarding/ReviewAndSubmitScreen';
 import MessagesStackNavigator from './messagesStackNavigator';
+import MainNavigationStack from './mainNavigationStack';
+import AccountNavigationStack from './accountNavigationStack';
+import {useAuth} from '../context/AuthContext';
 
 const Tab = createBottomTabNavigator();
-const Stack = createNativeStackNavigator();
-
-function HomeStack() {
-  return (
-    <Stack.Navigator
-      screenOptions={{
-        headerShown: false,
-      }}>
-      <Stack.Screen name="Home" component={HomeScreen} />
-      <Stack.Screen name="LoginScreen" component={LoginScreen} />
-      <Stack.Screen name="SignupScreen" component={SignupScreen} />
-      <Stack.Screen name="GuestDashboard" component={GuestDashboard} />
-      <Stack.Screen name="Scan" component={Scan} />
-      <Stack.Screen name="Pay" component={Pay} />
-      <Stack.Screen name="Bookings" component={Bookings} />
-      <Stack.Screen name="Pocket" component={Pocket} />
-      <Stack.Screen name="Detailpage" component={Detailpage} />
-      <Stack.Screen name="onBoarding1" component={onBoarding1} />
-      <Stack.Screen
-        name="personalDetailsForm"
-        component={personalDetailsForm}
-      />
-      <Stack.Screen
-        name="finalBookingOverview"
-        component={finalBookingOverview}
-      />
-      <Stack.Screen name="simulateStripe" component={simulateStripe} />
-      <Stack.Screen name="paymentAccepted" component={paymentAccepted} />
-      <Stack.Screen name="paymentDeclined" component={paymentDeclined} />
-      <Stack.Screen
-        name="bookedAccommodation"
-        component={bookedAccommodation}
-      />
-      <Stack.Screen name="Landing" component={Landing} />
-      <Stack.Screen name="Profile" component={profile} />
-      <Stack.Screen name="PaymentMethods" component={paymentMethods} />
-      <Stack.Screen name="Reviews" component={reviews} />
-      <Stack.Screen name="Settings" component={settings} />
-      <Stack.Screen name="HelpAndFeedback" component={helpAndFeedback} />
-      {/*<Stack.Screen name="HostHomepage" component={HostHomepage} />*/}
-      <Stack.Screen name="HostDashboard" component={HostDashboard} />
-      <Stack.Screen name="HostCalendar" component={HostCalendar} />
-      <Stack.Screen name="HostReviews" component={HostReviews} />
-      <Stack.Screen name="HostProfile" component={HostProfile} />
-      <Stack.Screen name="OnboardingHost" component={OnboardingHost} />
-      <Stack.Screen name="HostPayments" component={HostPayments} />
-      <Stack.Screen name="HostListings" component={HostListings} />
-      <Stack.Screen name="HostSettings" component={HostSettings} />
-      <Stack.Screen name="emailSettings" component={emailSettings} />
-      {/*<Stack.Screen name="HostRevenue" component={HostRevenue} />*/}
-      <Stack.Screen name="ListProperty" component={ListProperty} />
-      {/*<Stack.Screen name="LocationFillIn" component={LocationFillIn} />*/}
-      {/*<Stack.Screen name="selectAmenities" component={selectAmenities} />*/}
-      {/*<Stack.Screen name="PriceProperty" component={PriceProperty} />*/}
-      <Stack.Screen name="Register" component={Register} />
-      <Stack.Screen name="ConfirmEmail" component={ConfirmMail} />
-      <Stack.Screen name="HostDetailPage" component={HostDetailPage} />
-      {/*<Stack.Screen*/}
-      {/*  name="ReviewAndSubmitScreen"*/}
-      {/*  component={ReviewAndSubmitScreen}*/}
-      {/*/>*/}
-    </Stack.Navigator>
-  );
-}
 
 function BottomTabNavigator() {
+  const {isAuthenticated, user, userAttributes, checkAuth} = useAuth();
+
   return (
     <Tab.Navigator
       screenOptions={({route}) => ({
@@ -158,15 +55,27 @@ function BottomTabNavigator() {
       })}>
       <Tab.Screen
         name="Home"
-        component={HomeStack}
-        options={{headerShown: false}}
+        component={MainNavigationStack}
+        options={{
+          headerShown: false,
+        }}
       />
+      {isAuthenticated && (
+        <Tab.Screen
+          name="Messages"
+          component={MessagesStackNavigator}
+          options={{
+          headerShown: false,
+          }}
+        />
+      )}
       <Tab.Screen
-        name="Messages"
-        component={MessagesStackNavigator}
-        options={{headerShown: false}}
+        name="Account"
+        component={AccountNavigationStack}
+        options={{
+          headerShown: false,
+        }}
       />
-      <Tab.Screen name="Account" component={AccountPage} />
     </Tab.Navigator>
   );
 }
