+++ app/Domits/src/navigation/messagesStackNavigator.js
@@ -60,7 +60,7 @@ const MessagesStackNavigator = () => {
       }}
     >
       <Stack.Screen
-        name="Messages"
+        name="MessagesScreen"
         component={MessagesTab}
         options={{
           headerTitle: '',
